ImporterContextのAssetImport向けの機能を、EditorImporterContextに分離
EditorImporterContext
#758
