+++ Assets/UniGLTF/Editor/EditorImporterContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1c496f43ae12f9041a16af0489727587
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
