+++ Assets/UniGLTF/Runtime/UniGLTF/IO/ImporterContext.cs
@@ -2,11 +2,7 @@
 using System.Linq;
 using System.Collections.Generic;
 using UnityEngine;
-using System.IO;
 using UniGLTF.AltTask;
-#if UNITY_EDITOR
-using UnityEditor;
-#endif
 
 namespace UniGLTF
 {
@@ -230,237 +226,22 @@ namespace UniGLTF
         public List<AnimationClip> AnimationClips = new List<AnimationClip>();
         #endregion
 
-        protected virtual IEnumerable<UnityEngine.Object> ObjectsForSubAsset()
-        {
-            foreach (var x in TextureFactory.ObjectsForSubAsset())
-            {
-                yield return x;
-            }
-            foreach (var x in MaterialFactory.ObjectsForSubAsset())
-            {
-                yield return x;
-            }
-            foreach (var x in Meshes) { yield return x.Mesh; }
-            foreach (var x in AnimationClips) { yield return x; }
-        }
-
-#if UNITY_EDITOR
-        #region Assets
-        public bool MeshAsSubAsset = false;
-
-        protected virtual UnityPath GetAssetPath(UnityPath prefabPath, UnityEngine.Object o)
-        {
-            if (o is Material)
-            {
-                var materialDir = prefabPath.GetAssetFolder(".Materials");
-                var materialPath = materialDir.Child(o.name.EscapeFilePath() + ".asset");
-                return materialPath;
-            }
-            else if (o is Texture2D)
-            {
-                var textureDir = prefabPath.GetAssetFolder(".Textures");
-                var texturePath = textureDir.Child(o.name.EscapeFilePath() + ".asset");
-                return texturePath;
-            }
-            else if (o is Mesh && !MeshAsSubAsset)
-            {
-                var meshDir = prefabPath.GetAssetFolder(".Meshes");
-                var meshPath = meshDir.Child(o.name.EscapeFilePath() + ".asset");
-                return meshPath;
-            }
-            else
-            {
-                return default(UnityPath);
-            }
-        }
-
-        public virtual bool AvoidOverwriteAndLoad(UnityPath assetPath, UnityEngine.Object o)
-        {
-            if (o is Material)
-            {
-                var loaded = assetPath.LoadAsset<Material>();
-
-                // replace component reference
-                foreach (var mesh in Meshes)
-                {
-                    foreach (var r in mesh.Renderers)
-                    {
-                        for (int i = 0; i < r.sharedMaterials.Length; ++i)
-                        {
-                            if (r.sharedMaterials.Contains(o))
-                            {
-                                r.sharedMaterials = r.sharedMaterials.Select(x => x == o ? loaded : x).ToArray();
-                            }
-                        }
-                    }
-                }
-
-                return true;
-            }
-
-            return false;
-        }
-
-        public void SaveAsAsset(UnityPath prefabPath)
-        {
-            ShowMeshes();
-
-            //var prefabPath = PrefabPath;
-            if (prefabPath.IsFileExists)
-            {
-                // clear SubAssets
-                foreach (var x in prefabPath.GetSubAssets().Where(x => !(x is GameObject) && !(x is Component)))
-                {
-                    GameObject.DestroyImmediate(x, true);
-                }
-            }
-
-            //
-            // save sub assets
-            //
-            var paths = new List<UnityPath>(){
-                prefabPath
-            };
-            foreach (var o in ObjectsForSubAsset())
-            {
-                if (o == null) continue;
-
-                var assetPath = GetAssetPath(prefabPath, o);
-                if (!assetPath.IsNull)
-                {
-                    if (assetPath.IsFileExists)
-                    {
-                        if (AvoidOverwriteAndLoad(assetPath, o))
-                        {
-                            // 上書きせずに既存のアセットからロードして置き換えた
-                            continue;
-                        }
-                    }
-
-                    // アセットとして書き込む
-                    assetPath.Parent.EnsureFolder();
-                    assetPath.CreateAsset(o);
-                    paths.Add(assetPath);
-                }
-                else
-                {
-                    // save as subasset
-                    prefabPath.AddObjectToAsset(o);
-                }
-            }
-
-            // Create or update Main Asset
-            if (prefabPath.IsFileExists)
-            {
-                Debug.LogFormat("replace prefab: {0}", prefabPath);
-                var prefab = prefabPath.LoadAsset<GameObject>();
-#if UNITY_2018_3_OR_NEWER
-                PrefabUtility.SaveAsPrefabAssetAndConnect(Root, prefabPath.Value, InteractionMode.AutomatedAction);
-#else
-                PrefabUtility.ReplacePrefab(Root, prefab, ReplacePrefabOptions.ReplaceNameBased);
-#endif
-
-            }
-            else
-            {
-                Debug.LogFormat("create prefab: {0}", prefabPath);
-#if UNITY_2018_3_OR_NEWER
-                PrefabUtility.SaveAsPrefabAssetAndConnect(Root, prefabPath.Value, InteractionMode.AutomatedAction);
-#else
-                PrefabUtility.CreatePrefab(prefabPath.Value, Root);
-#endif
-            }
-            foreach (var x in paths)
-            {
-                x.ImportAsset();
-            }
-        }
-
-        /// <summary>
-        /// Extract images from glb or gltf out of Assets folder.
-        /// </summary>
-        /// <param name="prefabPath"></param>
-        public void ExtractImages(UnityPath prefabPath)
-        {
-            var prefabParentDir = prefabPath.Parent;
-
-            // glb buffer
-            var folder = prefabPath.GetAssetFolder(".Textures");
-
-            //
-            // https://answers.unity.com/questions/647615/how-to-update-import-settings-for-newly-created-as.html
-            //
-            int created = 0;
-            for (int i = 0; i < GLTF.textures.Count; ++i)
-            {
-                folder.EnsureFolder();
-
-                var gltfTexture = GLTF.textures[i];
-                var gltfImage = GLTF.images[gltfTexture.source];
-                var src = Storage.GetPath(gltfImage.uri);
-                if (UnityPath.FromFullpath(src).IsUnderAssetsFolder)
-                {
-                    // asset is exists.
-                }
-                else
-                {
-                    var byteSegment = GLTF.GetImageBytes(Storage, gltfTexture.source);
-                    var textureName = gltfTexture.name;
-
-                    // path
-                    var dst = folder.Child(textureName + gltfImage.GetExt());
-                    File.WriteAllBytes(dst.FullPath, byteSegment.ToArray());
-                    dst.ImportAsset();
-
-                    // make relative path from PrefabParentDir
-                    gltfImage.uri = dst.Value.Substring(prefabParentDir.Value.Length + 1);
-                    ++created;
-                }
-            }
-
-            if (created > 0)
-            {
-                AssetDatabase.Refresh();
-            }
-
-            // texture will load from assets
-            m_textureFactory.ImageBaseDir = prefabParentDir;
-        }
-        #endregion
-#endif
-
-        /// <summary>
-        /// Destroy the GameObject that became the basis of Prefab
-        /// </summary>
-        public void EditorDestroyRoot()
-        {
-            if (Root != null) GameObject.DestroyImmediate(Root);
-        }
-
-        /// <summary>
-        /// Destroy assets that created ImporterContext. This function is clean up for importer error.
-        /// </summary>
-        public void EditorDestroyRootAndAssets()
-        {
-            // Remove hierarchy
-            if (Root != null) GameObject.DestroyImmediate(Root);
-
-            // Remove resources. materials, textures meshes etc...
-            foreach (var o in ObjectsForSubAsset())
-            {
-                UnityEngine.Object.DestroyImmediate(o, true);
-            }
-        }
-
         /// <summary>
         /// Importに使った一時オブジェクトを破棄する
         /// 
         /// 変換のあるテクスチャで、変換前のもの
         /// normal, occlusion, metallicRoughness
         /// </summary>
-        public void Dispose()
+        public virtual void Dispose()
         {
-            // m_textureFactory.Dispose();
+            foreach (var x in m_textureFactory.Textures)
+            {
+                if (!x.IsUsed)
+                {
+                    // Destroy temporary texture object
+                    UnityEngine.Object.Destroy(x.Texture);
+                }
+            }
         }
 
         /// <summary>
