+++ Assets/VRM/Editor/Format/VRMEditorImporterContext.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58b502cc8f926b74fadb536f3b033100
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
