+++ Assets/VRM/Runtime/IO/VRMImporterContext.cs
@@ -298,81 +298,8 @@ namespace VRM
 
             return meta;
         }
-
-        protected override IEnumerable<UnityEngine.Object> ObjectsForSubAsset()
-        {
-            foreach (var x in base.ObjectsForSubAsset())
-            {
-                yield return x;
-            }
-
-            yield return AvatarDescription;
-            yield return HumanoidAvatar;
-
-            if (BlendShapeAvatar != null && BlendShapeAvatar.Clips != null)
-            {
-                foreach (var x in BlendShapeAvatar.Clips)
-                {
-                    yield return x;
-                }
-            }
-            yield return BlendShapeAvatar;
-
-            yield return Meta;
-        }
-
-#if UNITY_EDITOR
-        public override bool AvoidOverwriteAndLoad(UnityPath assetPath, UnityEngine.Object o)
-        {
-            if (o is BlendShapeAvatar)
-            {
-                var loaded = assetPath.LoadAsset<BlendShapeAvatar>();
-                var proxy = Root.GetComponent<VRMBlendShapeProxy>();
-                proxy.BlendShapeAvatar = loaded;
-
-                return true;
-            }
-
-            if (o is BlendShapeClip)
-            {
-                return true;
-            }
-
-            return base.AvoidOverwriteAndLoad(assetPath, o);
-        }
-
-        protected override UnityPath GetAssetPath(UnityPath prefabPath, UnityEngine.Object o)
-        {
-            if (o is BlendShapeAvatar
-                || o is BlendShapeClip)
-            {
-                var dir = prefabPath.GetAssetFolder(".BlendShapes");
-                var assetPath = dir.Child(o.name.EscapeFilePath() + ".asset");
-                return assetPath;
-            }
-            else if (o is Avatar)
-            {
-                var dir = prefabPath.GetAssetFolder(".Avatar");
-                var assetPath = dir.Child(o.name.EscapeFilePath() + ".asset");
-                return assetPath;
-            }
-            else if (o is VRMMetaObject)
-            {
-                var dir = prefabPath.GetAssetFolder(".MetaObject");
-                var assetPath = dir.Child(o.name.EscapeFilePath() + ".asset");
-                return assetPath;
-            }
-            else if (o is UniHumanoid.AvatarDescription)
-            {
-                var dir = prefabPath.GetAssetFolder(".AvatarDescription");
-                var assetPath = dir.Child(o.name.EscapeFilePath() + ".asset");
-                return assetPath;
-            }
-            else
-            {
-                return base.GetAssetPath(prefabPath, o);
-            }
-        }
-#endif
     }
+
+
+
 }
